backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
playwright==1.47.0
beautifulsoup4==4.12.3
httpx==0.27.0
pillow==10.3.0
pyzbar==0.1.9
python-dotenv==1.0.1
openai==1.51.2